estimated efforts added in task

+++ database/migrations/2025_10_01_112522_create_tasks_table.php
@@ -16,19 +16,21 @@ return new class extends Migration
             return;
         }
 
-        Schema::create('tasks', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('customer_id');
-            $table->unsignedBigInteger('user_id');
-            $table->text('task');
-            $table->unsignedBigInteger('created_by');
-            $table->timestamps();
+        if (!Schema::hasTable('tasks')) {
+            Schema::create('tasks', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('customer_id');
+                $table->unsignedBigInteger('user_id');
+                $table->text('task');
+                $table->unsignedBigInteger('created_by');
+                $table->timestamps();
 
-            // Foreign keys
-            $table->foreign('customer_id')->references('id')->on('customers')->onDelete('cascade');
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->foreign('created_by')->references('id')->on('users')->onDelete('cascade');
-        });
+                // Foreign keys
+                $table->foreign('customer_id')->references('id')->on('customers')->onDelete('cascade');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->foreign('created_by')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
     }
 
     /**
